refactor: unify auth service
