+++ src/ai_karen_engine/api_routes/auth.py
@@ -15,7 +15,7 @@ from ai_karen_engine.core.dependencies import (
 )
 from ai_karen_engine.core.chat_memory_config import settings
 from ai_karen_engine.core.logging import get_logger
-from ai_karen_engine.services.unified_auth_service import auth_service
+from ai_karen_engine.security.auth_service import auth_service
 
 logger = get_logger(__name__)
 router = APIRouter(tags=["auth"])
