+++ src/ai_karen_engine/services/auth_utils.py
@@ -3,7 +3,7 @@ from typing import Any, Dict, Optional
 from fastapi import HTTPException, Request, Response, status
 
 from ai_karen_engine.core.chat_memory_config import settings
-from ai_karen_engine.services.auth_service import auth_service
+from ai_karen_engine.security.auth_service import auth_service
 
 # Shared cookie name for authentication sessions
 COOKIE_NAME = "kari_session"
