+++ src/ai_karen_engine/utils/auth.py
@@ -9,7 +9,7 @@ import asyncio
 
 import jwt
 
-from ai_karen_engine.services.auth_service import auth_service
+from ai_karen_engine.security.auth_service import auth_service
 from ai_karen_engine.core.logging import get_logger
 
 logger = get_logger(__name__)
